// src/app/_dto/bankdetails.dto.ts
export class bankdetailsdto{
  HEBank_Name: string;
  HEBank_Acc_No: string;
  HEBank_Add1_No: string;
  HEBank_Add2_No: string;
  HEBank_Add3_No: string;
  HEBank_Add4_No: string;
  HEBank_City_SysID: string;
  HEBank_City_Code: string;
  HEBank_City_Name: string;
  HEBank_Count_SysID: string;
  HEBank_Count_Code: string;
  HEBank_Count_Name: string;
  HEBank_Tele_No: string;
  HEBank_Fax_No: string;
  HEBank_Email: string;
  HEBank_IBANAcc_No: string;
  HEBank_RouAcc_No: string;
  HEBank_Swift_No: string;
  HEBank_Oth1_No: string;
  HEBank_Oth2_No: string;
  HEBank_Oth3_No: string;
  HEBank_Oth4_No: string;
}