// src/app/_dto/emergencydetails.dto.ts
export class emergencydetails{
    HEmerAdd_First_Name: number;
    HEmerAdd_Sec_Name: string;
    HEmerAdd_Full_Name: string;
    HEmerAdd_Add1_Name: string;
    HEmerAdd_Add2_Name: string;
    HEmerAdd_Add3_Name: string;
    HEmerAdd_Add4_Name: string;
    HEmerAdd_City_Code: string;
    HEmerAdd_City_Name: string;
    HEmerAdd_Country_Code: string;
    HEmerAdd_Country_Name: string;
    HEmerAdd_Tel_Number: string;
    HEmerAdd_Fax_Number: string;
    HEmerAdd_Email: string;
}