// src/app/_dto/emiratedetails.dto.ts
export class emiratedetails{
    HEID_Number: string;  // Assuming it's a varchar (string)
    HEID_Bdate: Date;  
    HEID_Dissue: Date;  
    HEID_Validtill: Date;  
    HEID_City_SysID: number;  
    HEID_City_Code: string;  
    HEID_City_Name: string;  
    HEID_Count_SysID: number;  
    HEID_Count_Code: string;  
    HEID_Count_Name: string;  
    HEID_Place_Issue: string;  
    HEID_Mstatus: string;  
}