// src/app/_dto/permenantaddress.dto.ts
export class permanentaddressdto{
    HEPAdd_First_Name: string;
    HEPAdd_Sec_Name: string;
    HEPAdd_Full_Name: string;
    HEPAdd_Add1_Name: string;
    HEPAdd_Add2_Name: string;
    HEPAdd_Add3_Name: string;
    HEPAdd_Add4_Name: string;
    HEPAdd_City_Code: string;
    HEPAdd_City_Name: string;
    HEPAdd_Country_Code: string;
    HEPAdd_Country_Name: string;
    HEPAdd_Tel_Number: string;
    HEPAdd_Fax_Number: string;
    HEPAdd_Email: string;

}