// src/app/_dto/presentaddress.dto.ts
export class presentaddressdto {
  HEAdd_First_Name: number;
  HEAdd_Sec_Name: string;
  HEAdd_Full_Name: string;
  HEAdd_Add1_Name: string;
  HEAdd_Add2_Name: string;
  HEAdd_Add3_Name: string;
  HEAdd_Add4_Name: string;
  HEAdd_City_Code: string;
  HEAdd_City_Name: string;
  HEAdd_Country_Code: string;
  HEAdd_Country_Name: string;
  HEAdd_Tel_Number: string;
  HEAdd_Fax_Number: string;
  HEAdd_Email: string;
}   