// src/app/_dto/TbldLeaveSum.dto.ts
export class TbldLeaveSum{
    DLSum_SysID: number;
    DLSum_Emp_SysID: number;
    DLSum_Emp_Code: string;
    DLSum_Emp_Name: string;
    DLSum_Join_Date: Date;
    DLSum_Term_Date: Date;
    DLSum_LastWork_Date: Date;
    DLSum_Calcu_Date: Date;
    DLSum_Lsal_Days: number;
    DLSum_LsalUsed_Days: number;
    DLSum_LsalBal_Days: number;
    DLSum_Gratu_Days: number;
    DLSum_GratuUsed_Days: number;
    DLSum_GratuBal_Days: number;
    DLSum_Absen_Days: number;
    DLSum_Emplo_Status: string;
    selected :boolean;
}